+++ tests/Service/Connecteur/ConnecteurAssociationServiceTest.php
@@ -6,8 +6,6 @@ use Exception;
 use FluxEntiteSQL;
 use Pastell\Service\Connecteur\ConnecteurActionService;
 use Pastell\Service\Connecteur\ConnecteurAssociationService;
-use Pastell\Service\FeatureToggle\DisplayConnecteurEntiteRacine;
-use Pastell\Service\FeatureToggleService;
 use PastellTestCase;
 use UnrecoverableException;
 
@@ -131,10 +129,6 @@ class ConnecteurAssociationServiceTest extends PastellTestCase
      */
     public function testHeritageConnecteurEntiteRacine(): void
     {
-        /** @deprecated 4.1.7, to be removed in v5 */
-        $this->getObjectInstancier()
-            ->getInstance(FeatureToggleService::class)
-            ->enable(DisplayConnecteurEntiteRacine::class);
 
         $id_ce = $this->createConnector(
             'transformation-generique',